# main.go
package main

import (
	"blogpost/clients"
	"blogpost/config"
	docs "blogpost/docs" // docs is generated by Swag CLI, you have to import it.
	"blogpost/handlers"
	"log"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// @license.name Apache 2.0
// @license.url  http://www.apache.org/licenses/LICENSE-2.0.html
func main() {
	conf := config.Load()

	// programmatically set swagger info
	docs.SwaggerInfo.Title = conf.App
	docs.SwaggerInfo.Version = conf.AppVersion

	router := gin.New()

	if conf.Environment != "development" {
		gin.SetMode(gin.ReleaseMode)
	}

	if conf.Environment != "production" {
		router.Use(gin.Logger(), gin.Recovery()) // Later they will be replaced by custom Logger and Recovery
	}

	grpcClients, err := clients.NewGrpcClients(conf)

	if err != nil {
		panic(err)
	}

	defer grpcClients.Close()

	h := handlers.NewHandler(conf, grpcClients)

	v1 := router.Group("/v2")
	{
		v1.Use(MyCORSMiddleware())
		v1.POST("/login", h.Login)

		v1.POST("/article", h.AuthMiddleware("*"), h.CreateArticle)
		v1.GET("/article/:id", h.AuthMiddleware("*"), h.GetArticleByID)
		v1.GET("/article", h.AuthMiddleware("*"), h.GetArticleList)
		v1.PUT("/article", h.AuthMiddleware("*"), h.UpdateArticle)
		v1.DELETE("/article/:id", h.AuthMiddleware("*"), h.DeleteArticle)

		v1.POST("/author", h.AuthMiddleware("*"), h.CreateAuthor)
		v1.GET("/author/:id", h.AuthMiddleware("*"), h.GetAuthorByID)
		v1.GET("/author", h.AuthMiddleware("*"), h.GetAuthorList)
		v1.PUT("/author", h.AuthMiddleware("*"), h.UpdateAuthor)
		v1.DELETE("/author/:id", h.AuthMiddleware("*"), h.DeleteAuthor)
	}

	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	router.Run(conf.HTTPPort) // listen and serve on 0.0.0.0:8080 (for windows "localhost:8080")
}

//MyCORSMiddleware ...
func MyCORSMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		log.Println("MyCORSMiddleware...")

		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Credentials", "true")
		c.Header("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, PATCH, DELETE")
		c.Header("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Header("Access-Control-Max-Age", "3600")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	}
}